excercise 8 global objects
1.	Using the JavaScript examples on pages 128 – 129 as a reference, 
repeat the series of property and 
method requests for a string of 5 or more words.

2.	Use the Number object to 
a.	Determine if “abc” is not a number
b.	Round 23.98673495 to 3 decimal places
c.	Limit 765.7898 to 1 decimal place

3.	Use the Math object to 
a.	Round 143.436
b.	Derive the square root of 192
c.	Round 45.23 up to the nearest integer
d.	Round 38.91 down to the nearest integer
e.	Create a random number between 0 and 100.

4.	Using the Date object
a.	Create a date for the 2nd of April 1999
b.	Retrieve the value of minutes currently